--- app/Http/Controllers/API/ContactAPIController.php
@@ -1,132 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\API;
-
-use App\Http\Controllers\Controller;
-use App\Models\Contact;
-use App\Models\WebUser;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\Facades\DB;
-
-class ContactAPIController extends Controller
-{
-
-    //rule to show error
-    public function ruleValidates()
-    {
-
-        return [
-            'name' => 'required',
-            'email' => 'required|email',
-            'phone' => 'required|regex:/(0)[0-9]{9}/',
-            'description' => 'required',
-        ];
-    }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-        $validation = Validator::make(
-            $request->all(),
-            $this->ruleValidates()
-        );
-        if ($validation->fails()) {
-            $response = array('status' => 'error', 'errors' => $validation->errors()->toArray());
-            return response()->json($response);
-        }
-
-        $user = new WebUser();
-        $user->name = $request->name;
-        $user->phone = $request->phone;
-        $user->mail = $request->email;
-
-        $user->save();
-
-        $contact = new Contact();
-
-        $contact->user_id = DB::table('users')
-            ->where('name', '=', $request->name)
-            ->where('phone', '=', $request->phone)
-            ->where('mail', '=', $request->email)->value('id');
-
-        $contact->save();
-
-        if ($user && $contact) {
-            return response()->json(["status" => 200, "success" => true, "message" => "Thank for your submit we will connect you soon"]);
-        } else {
-            return response()->json(["status" => "failed", "success" => false, "message" => "Whoops! failed to submit."]);
-        }
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
-    }
-}
