+++ app/Http/Controllers/Api/DetailSessionAPIController.php
