--- app/Models/WebUser.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class WebUser extends Model
-{
-    use HasFactory;
-    protected $table= "users";
-}
